requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
sympy
streamlit
Pillow
datawhispers